+++ src/controllers/TaxZonesController.php
@@ -62,7 +62,7 @@ class TaxZonesController extends BaseTaxSettingsController
             $variables['title'] = Craft::t('commerce', 'Create a tax zone');
         }
 
-        $countries = Plugin::getInstance()->getCountries()->getAllCountries();
+        $countries = Plugin::getInstance()->getCountries()->getAllEnabledCountries();
         $states = Plugin::getInstance()->getStates()->getAllStates();
 
         $variables['countries'] = ArrayHelper::map($countries, 'id', 'name');